[Improvement-13653][Remote Logging] Support api-server to get task log from remote target
